requirements.txt
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
requests
psycopg2-binary
# psycopg
# psycopg2